move wildcard views
