+++ examples/todomvc/src/components/App.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Header from '../containers/Header'
 import MainSection from '../containers/MainSection'
-import WcTable from '../containers/WcTable'
+import WcTable from '../wildcard/WcTableContainer'
 
 const App = () => (
   <div>
