+++ examples/todomvc/src/wildcard/WcTableComponent.js
