+++ examples/todomvc/src/wildcard/WcTableContainer.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux'
 import * as TodoActions from '../actions'
 import * as WildcardActions from '../wildcard/actions'
 import { bindActionCreators } from 'redux'
-import WcTable from '../components/WcTable'
+import WcTable from './WcTableComponent'
 import { getSortedTodos } from '../selectors'
 
 const mapStateToProps = state => ({
